+++ NEWS.md
@@ -1,10 +1,12 @@
 # COVIDutilities 1.1.0
-*remove COVID related urls from exports
+* change default option to `get_vax_countries` to only return PEPFAR countries.
+* `get_vax_data` converts country names to match PEPFAR ones where applicable. 
+* clean up exports, removing COVID related urls and background supporting functions.
 
 # COVIDutilities 1.0.0
-* add `pull_stringency_index()` to pull Oxford stringency data
+* add `pull_stringency_index()` to pull Oxford stringency data.
 * add `get_vax_data()`  to fetch COVID-19 vaccine data from Our World In Data Github Repo.
-* add parameter to `pull_jhu()` to limit to just PEPFAR countries
+* add parameter to `pull_jhu()` to limit to just PEPFAR countries.
 
 # COVIDutilities 0.0.0.9000